Handle pull status
