+++ lib/util/now_displaying_manager.dart
@@ -158,21 +158,4 @@ class NowDisplayingManager {
     );
     return res;
   }
-
-  Timer? _statusPullTimer;
-
-  void startStatusPull() {
-    _statusPullTimer?.cancel();
-    _statusPullTimer = Timer.periodic(
-      const Duration(seconds: 5),
-      (timer) async {
-        await updateDisplayingNow(addStatusOnError: false);
-      },
-    );
-  }
-
-  void stopStatusPull() {
-    _statusPullTimer?.cancel();
-    _statusPullTimer = null;
-  }
 }
